requirements.txt
streamlit==1.31.1
streamlit-image-coordinates==0.1.6
Pillow==9.3.0
matplotlib==3.7.1
numpy==1.24.1
segment-anything==1.0
torch==2.0.1